Fixed exception in jdbc read from new fly servers, 4-1 & 3-2. for target features the organism in fly is always the dummy "Computation Result". This used to always have the abbreviation "Comp". Now theres 2 computational result organisms (which is rather suspicious) and one has the abbreviation and one doesnt. The one that doesnt is being used and was causing the jdbc reader to choke. The jdbc reader was using the organism abbreviation to name the feature set. Since it was null, this caused a null pointer exception from Range. _makeFeatureSet now checks for null, and uses RangeI.NO_NAME in its place.
maybe Range should just set name to NO_NAME when its null
 rather than throwing the exception?
or is this purposeful to let the dataadapter know something is
wrong. Range could be tolerating but print out a warning message to
 stdout in that case, that would be a little more robust wouldnt it?
Rather than throwing a null pointer which usually completely kills a
 data adapter load.

So also in the case that the organism abbreviation is null, the
organism full name, genus+species, is queried for in its place.