i changed how update parent is handled for exons. previoulsy annotation editor moveSingleExonToTrans would create a update parent event. TransactionManager would then convert this into a delete and add (guanming had to do this for rice - forget why). this would mean if you wanted to track then an exon had moved from one trans to another that info would be lost, and in fact tair does want to track this.
so now an UpdateParentTransaction is created, that is a compound
transaction with delete & add trans as children. a data adapter can do
whatever it likes with this trans - a move or a del/add. the del/add
are convenient for coalescing, as well as if yer not dealing with
moves.

annotation change events should eventually become just wrappers of
transactions as they are redundant - getting closer to that -
ACE.getParentFeature now checks parent of transaction.

merge exons was sending out a change limits transaction even if the
limits of the feat hadnt actually changed - now checks for that.

moveSingleExonToTrans was doing exon overlap check AFTER it had added
the exon to the trans - which was wrong as it would then always find
an overlap with itself - rather silly - it now looks for overlaps
before it adds the exon.

a compound trans can indicate it doesnt want to get flattened on being
added to a parent compound trans, so it doesnt get wiped out, as is
the case with update parent - dont want to lose the move info.

drawable repairFeatureSet can now deal with compound events - just
recurses them to leaves.