You can now layer GAME data (results or annots). The first GAME file you load needs to have a <seq> block with focus=true.  The second one can have <seq> or not.
Also removed exon sanitizing that we put in a while back to fix
bad exon types that were created by an old version of Apollo with a
bug in printing exon types.  Hopefully there are no more old buggy
datafiles like that sitting around.  Forcing exon types to be "exon"
is dangerous in case in the future the data has features at the exon
level that are not exons.