with the new much bigger rice database, inserting data ground to a halt - it was taking over 10 minutes to insert a simple little gene. theres 2 issues here - the main one is select max id from feature - to find out what id got inserted. the max id query takes over a minute and is done for every feature (gene,trans,prot,ex). its really odd that it takes so long as there is an index on it so you would think it would just shoot to the end of the index.
so i put in the nextval function which figures the next id to be inserted.
and then does the insert with this id. the bummer is that i imagine
nextval is postgres specific - this will need to be configured when
other vendors want to use writeback.

the other slowness is feat_rel insert which must be a trigger issue -
scott is looking into it.