Fixed problem reported by Haiyan Zhang: Apollo changed the genes' dbxref->db_id from FlyBase to GadFly (was actually hardcoded before--now it's recorded in the dbxref).
Also now saving the genomic db in the dbxref for the curation set--this
comes from the result queries.  Saving the one from the first result
query we get to and assuming it's the same for subsequent result
queries--saving each one would waste too much space, and checking that
each one is in fact the same would waste time.