These are from Chris Wilks of TAIR actually. I think in the future he will just commit directly. Heres his comments:
+SeqFeatureI and SeqFeature
added user object stuff
This allows for ORM layers (lie tair) to have a hook into SeqFeatures.

+CompoundTransaction
changed protected hasCompoundType() to public

The goal is to get tair synched up with the latest apollo, so they
dont have to either stay with an old version, or keep re-adding there
changes with every update.