added sending selection from apollo to igb. its rather basic at the moment - proof of concept. this is configured in style file with
EnableIgbHttpConnection "true" - default is false

Also this only works in DEBG at the moment. at the moment its blindly
sending out the url to localhost - i would like to add stuff to detect
if igb is actually around (check different a few ports like igb does)

theres a new external package that has the new IgbBridge. IgbBridge
listens for selections from controller and sends out http to igb when
it gets them. its just doing the region of the feature selected - is
there anyway else to go in igb? probably that is sufficient - might
want to pad it a bit.

gui curation state adds in igb bridge if the style is configured.

i made the package name "external" - im open to other suggestions here
- i contemplated "igb" - except at some point im gonna add the ability
to take in http requests which could come from igb but it could come
from somewhere else - so "igb" seemed not general enough. i suppose it
could be "http" - not sure if it will just be http stuff in this package.

im just using apollos HTMLUtil.loadIntoBrowser for now - eventually we
will want to be independent of a browser - no need to go through a
browser.

also its hardwired to lookup up drosophila in igb - need some organism
configuration. also need to get igb talking to apollo...