fixed an old bug that reared its head again - on doing a load from a different data adapter with different tiers the new tiers were failing to load - this used to happen with the TypesObserver constructor - i moved it guiCurationState calling firePropSchemeChangeEvent on the property scheme - to let the world know theres a new prop scheme. steve took this out on 8/8 - so this may cause problems with synteny - waiting to hear from him - may have to revisit this again to get it to work for both synteny and tier changes
since this bug has come up twice i tried to make a test case for
it. this was harder than i thought and i didnt end up succeeding -
with the prop scheme change event disabled a prop scheme event would
still get fired - in TestApollo the gui goes a little wack (not sure
why) and paints and layouts & excpetions seem to fly all over - so a
layout comes in and causes the prop scheme to update properly - which
doesnt happen in a normal apollo session - so i gave up on the test
for now - may revisit at another time.