Fixed bug noticed recently by Sima: Problem with stop codon not being recognized after gene is changed (seems to happen only when data is loaded from ChadoXML--same data loaded from GAME XML doesn't have this problem!)
For some reason, the fact that I was using GAMESequence, rather than Sequence,
in the ChadoXmlAdapter for the cDNAs and peptides, was the cause of the problem.
I'm not sure why; it seemed to have something to do with the GAMESequences not
updating properly when the transcript changed.  At any rate, there was no reason
to use GAMESequences; plain old Sequences seem to work just fine (though if we
notice any new problems with the ChadoXML adapter, we should remember to check
whether it could be an issue with the Sequences).