Can now print a whole SyntenyPanel. As PSGr2 does not implement Graphics2D only Graphics, this required me to modify the SyntenyLinkPolygen and SyntenyLinkPanel code which expected a Graphics2D to be able to handle either a Graphics2D or Graphics object.
Added some optimisations to Link drawing - don't draw outline polygon if
display is zoomed out. The drawPolygon call seemed to be the most
expensive part of the draw and when zoomed out it just make the display
look ugly anyway.

I also added some rendering hints to the Graphics2D to use the _SPEED
options for rendering, coloring and no antialiasing