Removed zscroll variable which wasn't used for anything
Added a controller.handleZoomEvent in zoomToWidth to allow link panels
in 3 (or more)-way synteny panels to be able to respond to zoom. Previously
this lead to incorrect display of links in one of the link panels when
the 'Equalize ranges' SyntenyLinkPanel menu item was selected in the
other link panel.

Added methods to allow the 'control panel' (panel with zoom buttons and
horizontal scrollbar) to be shown/hidden.

Removed a doLayout which should not be necessary now the
LinearView.setVisible method does one.