well this was a bit of a slippery slope - i thought it would be trivial to add location query for fly. i wanted it for 2 reasons - 1 this is a nice query to have in general and 2 its so friggin slow that if you can restrict the range to that which you absolutely need you can get a significant speedup - this will help me debugging a lot too as i wont have to bring up the 3 minute gene query over & over. anyways - so i configured chado-adapter.xml with "chromosome_arm" - well this not surprisingly did not create a gui with start & end - it turns out the chado adapter was hardwired to the term "chromosome" which is what rice uses for top level location queries - not so for fly which is chromosome_arm - so rather than just adding another term to the hardwiring i made it a configurable:
            <type>
                <name>chromosome</name>
                <useStartAndEnd>true</useStartAndEnd>
                <queryForValueList>true</queryForValueList>
            </type>

useStartAndEnd says to put in the start & end field. <type>gene</type>
is still valid for the common case where ya dont need to specify the
loc stuff (default start&end false, query false).

so i went one step further - queryForValueList actually queries the db
for the feat names with the type - which is kinda nice rather than
configging it and having it get out of synch (which has happened of
course) - had to put tests & exceptions in to test for connection.

im resusing the method getAllChadoSequencesByType (from
tigr/crabtree). but its more elaborate than is needed and is slow as
postgres chokes on the outer join - so i think im gonna add a new
method that just gets the names - (fly & rice dont have the
descriptions that tigr has)