crm/fly had new naming features - they wanted the annot name to just take on the result name. all the name adapters just did temp names for new names - even DefaultNameAdapter. so i made a SimpleNameAdapter that gives the annot the result name (should this be Default?).
stumbled on some annot editor bugs on the way - turns out on dragging
a new annot in an annot was being named with the name adapter for gene
- no matter what the name adapter was set to in the tiers file - this
was because the naming of the gene was happening before the setting of
the type. the setting of the top level type was occurring in
buildTranscript - which would see if all the results mapped to the
same annot type - and if not wouldnt set the type at all. this is
funny as well and i havent ammended it but need to revisit - i think
if there are results with different annot types then an exception be
thrown (or checked for earlier in the process) and in any case it
should be disallowed - what happens now is that too uses gene name
adapter and actually sets the type to "gene" (the default
AnnotatedFeature type) even if gene wasnt an annot type in the
results.

so there was one more problem - saving to chado with the same
uniquename as the result violates a unique constraint as they also
have the same type & organism - so i appended a "-annot" to the id in
SimpleNameAdapter generateName - this veers slightly from the
simpleness of the adapter but its no biggie, and can always be
configged or separated to a separate adapter at another point if we
need it.