followed up on some changes to overlap detection that nomi did and fixed bug sima was seeing - the bug was that it was asking about merges on tiling path fragments if there were 2 overlaps but not one (it shouldnt have been asking for either). also fixed a bug with comparing types.
so the bug fix is that AnnotationEditor.areOverlapping was comparing
features biotypes to see if they were the same - however one feat
could be a result - and it should be looking at the annot type that it
maps to - so it now does this - this means that a result dragged into
an orf of a gene now properly uses ORF_Overlap and creates a new gene
- before it was improperly using SimpleOverlap and adding as
transcript.

also areOverlapping now does the logic that was in addGeneOrTranscript
- if 2 annots are of different types then they dont overlap ever - im
pretty sure this is what we want - 2 different types should never
overlap and add themselves as transcripts.

NoOverlap has been added as a new OverlapI object - this is for 1
level annots that dont overlap even with each other - i believe this
is true that 1 level annots can be piled on top of each other and dont
merge into a new annot. i set 1 level annots in fly.tiers to use
NoOverlap

i think the rnas should be using SimpleOverlap and not ORF_Overlap as
they are now.

refactored type & biotype which i always found rather muddled & confusing -
now its featureType & topLevelType - biotype is just the type of your
top level feat