got apollo listening to http region request - albeit rather crudely. it presently doesnt do any checking - so it scrolls/zooms to the range given even if its way out of the current range, and also it doesnt check that the chromosome is the right one either. but proof of concept is there at least.
there isnt an event for setting a range, just for centering on a bp
with BaseFocusEvent - so i added the ability to do ranges in
BaseFocusEvent. seemed an ok use of BaseFocusEvent, but if this doesnt
seem right to others i can make a new RangeEvent or something like
that.

renamed getLinearVisibleWidth to getVisibleBasepairWidth for clarity.

this can be tested with something like:
http://localhost:8085/ApolloControl?seqid=chr2R&start=10000&end=12000
- i think ya hafta be in debug mode