fixed a memory leak in command line bulk loading. its a memory leak that was fixed for gui, but bulk loading was not using the fix/cleanup. AbstractLazySequence extended NonWindowControlledObject - and is the only class that extends it - NWCO adds itself as a listener to the controller - and the mem leak comes from not removing itself as a listener when new data is loaded (and seq has ref to cur set).
so i was about to put in the gui fix - where cleanup() is called and
its removed as a listener - but i looked a little deeper and saw in
fact theres no reason for ALS to be a listener to the controller - its
not listening for events and its interface ControlledObjectI is only
used by controller for listeners that have windows, which NCO/ALS does
not - so it seemed purposeless and it was causing mem leaks - so i
took out the extending of NWCO so its no longer listening to nothing,
which fixes the mem leak. Steve - do you know what function NWCO was
suppose to serve? now nothing is using it so if there is no use for it
it should be removed.

took out all the cleanup methods that are no longer needed.