Extended handling of stop_codon_readthrough: can now handle the specific stop_codon_redefinition_as_selenocysteine (which makes the stop codon a U) or the more generic stop_codon_readthrough.
The stop_codon_readthrough property can now have a boolean
(true/false) value or a character.  If it is true, the affected stop
codon will show up as an X.  If it is a character (e.g. "B"), it will
show up as that character in the Apollo display and also will be saved
that way in the ChadoXML file.  (This shouldn't screw anything up--the
only way it could be a weird character like that is if it were read in
that way, since you can't specify a particular readthrough stop codon
inside Apollo.)  The annotation info editor just has a checkbox for
"readthrough stop codon" (yes/no), which is checked if the
stop_codon_readthrough property is non-false and non-empty or the
stop_codon_redefinition_as_selenocysteine property is true--thus, you
can't tell the difference between 'U' and 'X' inside the annot info
editor.  Ideally, there'd be a text box for "stop codon readthrough
residue" or something, but that'd be a lot of work to add.

Another caveat is that although you can see the redefined stop codon (U,
X, or whatever) if you zoom in on the annotation, the Sequence window
just prints * for the readthrough stop codon.