Since I last looked at this code the following two things have changed (I think...):
1. It used to be the case that getConnection() would always return
   a new connection.  In the current code, however, getConnection()
   will return a cached copy of the Connection, if one exists.
2. There used to be only a single init() method and a single
   getConnection() method.  There are now two versions of each:
   a "withException" version that throws exceptions and the original
   version, which does not throw exceptions.

These changes both affect SybaseChadoAdapter, which was trying to
override getConnection() (to no avail, since
getConnectionWithException() is now doing the real work.)  To
fix this I've added a new method to JdbcChadoAdapter, getNewConnection(),
which implements the original semantics of the getConnection() method
and provides a handy target for overriding in a subclass.  The
documentation of the affected methods has been updated accordingly.
This should be a fully backwards-compatible change.