well this was a bit of a slippery slope - but apollo transactions can now be read from inside of game files - they no longer need to be saved as separate files - the style configuration for this is
TransactionXmlInGameFile "true"

i think the default is false - should it be true?

(i already got transactions saving to game file in last commit).

so the problem was that TransactionXMLAdapter uses org.w3c.dom to
build a DOM - the current java interface - but the game adapter is
funny - it uses sax but it builds a dom - and its not a org.w3c.dom -
its a homemade org.bdgp.xml.XMLElement dom (im guessing the reason for
this is perhaps the w3c interface was not around when the game adapter
came into existence) - so the game adapter has a XMLElement for the
transactions but the TransactionXMLAdapter which already knows how to
parse these buggers was choking on it as it wasnt a w3c Element -
yikes! so one lame solution woulda been to just build up 2 separate
doms adding more inefficiency to the already rather inefficient game
adapter (it should lose its dom and just use the underlying sax!) - or
redo all the xml parsing (perhaps go sax - a bit ambitious for the
task at hand) - or - and this is what i did - just make XMLElement
conform to the w3c interface so TransactionXMLAdapter can work with it
- and the whole xml file is not parsed twice. this meant putting in a
bunch of no-ops for w3c methods we dont use (a bunch went in with the
last commit) - this was made easier by having org.bdgp in apollo now.

also TansactionXMLAdapter now just adds transactions straight to the
TransactionManager which im tempted to rename TransactionList.