game can now read & write proteins as feature_spans under transcripts  feature_set:
      <feature_span id="RICE000402">
        <type>polypeptide</type>
        <name>RICE000402</name>
        <seq_relationship type="query" seq="chr10:46000-52000">
          <span>
            <start>157</start>
            <end>427</end>
          </span>
        </seq_relationship>
      </feature_span>

this allows for protein ids & names to be roundtripped - and i imagine
synonyms & dbxrefs could probably go in there as well - previously
protein id/names were not being roundtripped - this caused apollo to
think a changed protein was new and would try to insert an already
existing protein. in fact you cant make a new protein in apollo
independent of a new transcript or gene.

i guess prot dbxrefs could be displayed in dbxref table - no need for
prot syns as they are not displayed in apollo - unless you are doing
wipe out and reinsert and need to roundtrip unedited,unviewed
data. (and eventually apollo might do more with proteins)

this addition is a change to game and is part of game 1.1 - i need to
make the game version configurable (right now theres just a variable
in game save) - but is backwards compatible as it will still read in
either start_codon feature_span or the above polypeptide - i went with
polypeptide to be consistent with SO. not sure how to reconcile game
version configuring and 1 level annot configuring as they are
overlapping but not the same concepts.

refactored setTranslactionStartAtFirstCodon() to
calcTranslationStartForLongestPeptide() as it doesnt necasarily set
start at first codon if a longer orf is found at an existing start
codon.

oh also i fixed another bug - when get_cDNA was being called it was
wiping out the peptide seq but not setting it invalid - this led to an
empty peptide sequence - it used to just set peptide to null (another
form of invalidating) but that would lose peptide id & name. so now it
invalidates so the peptide is recalculated - i think this fixes andys
bug he reported the other day.

the protein seq feature is maturing - it has a funny relationship
presently with peptide sequence as thats where the id & name are
presently stored as thats where the rest of apollo code has always
fiddled with that stuff - eventually should be moved to prot feat i
think but it would be funny to leave pep seq with no name and id - and
its funny to have redundantly in both (synching issues) - and pep seq has
no pointer back to its feat where prot feat does have a pointer to pep
seq - whats unique about this situation is that theres a sequence
thats not irrelevant (like gene & transcript) and also who's name &
such get edited (unlike hit sequence). need to ruminate some more on
this.

refactored GAMEAdapter.getExon into getTranscriptFeatureSpan -
previously it was returning Exon objects even if it was a start_codon
span - now it return SeqFeatureI of Exon, Protein, and just
SeqFeatureI for start_codon (backward compatible).