Made ChadoXML adapter handle one-level annots correctly, instead of the kludge of promoting them to phony three-level annots and then collapsing them again for writing.
Amazingly, all that was needed to accomplish this was to comment out the
call to
    makeFakeTranscriptIfNeeded(annot, curation);
in ChadoXmlAdapter.  ChadoXmlWrite didn't need to be changed--if there are
no fake transcripts, it doesn't have to worry about not writing out fake
transcripts.

I commented out (but did not yet remove) a bunch of no-longer-needed code.