updated functions/feature_ids_fx.plpgsql to be compatible with pg8--now the question is whether they still work with pg7.
Changed apollo_split function to take three feature_ids and no long tries to create
new features in the function (as it isn't really possible).  So now the way to use this
is to create two empty genes, and then give them to the function which will
then transfer annotations from the old gene to the new genes.  Now I need to do the same
for the apollo_merge function.