If setTranslationEndFromStart finds that translation_start is not in the range of the feature set, it was trying to call calcTranslationStartForLongestPeptide() but this sometimes resulted in infinite loops where it'd try to set the translation start, then the end, then the start...
What I'm having it do now is that if translation start is not in the
range of the transcript, and missing_5prime is set (which it is, in
this case), then it doesn't try to calculate the correct translation
start or end--it just gives up.

That means that if there's some weird transcript with a weird translation
start value that *doesn't* have missing_5prime set to true, Apollo could
still loop, but I think that situation will be rare.