Slight refactoring that only affects TIGR databases.  At TIGR we distinguish (nominally, at least) between chado databases used for the annotation and/or analysis of a single organism (these are referred to as "annotation" or "project" databases) and chado databases used for the annotation and/or analysis of several related organisms (these are referred to as "comparative" or "Sybil" databases, because of the Sybil project at sybil.sourceforge.net.)
The method definitions formerly in TigrSybilChadoInstance have been moved
into TigrChadoInstance.  TigrSybilChadoInstance inherits from TigrChadoInstance
but is essentially empty; if it turns out that the distinction between
TIGR annotation and comparative databases is unnecessary then the subclass(es)
can be removed.  In the long term, of course, all the "*Instance" classes
should be replaced by sets of configurable parameters appearing in the
chado adapter config. files.