1. Renamed a couple of methods that were named inconsistently:   set/getHitPrograms -> set/getSearchHitPrograms   set/getPredictionPrograms -> set/getGenePredictionPrograms The new names match the corresponding XML tags in the config. file, which appears to be the convention used for most of the other setters and getters.
2. Added ChadoInstance.getSearchHitPrograms(); since the corresponding setter
is there and since both methods are defined in AbstractChadoInstance I'm
pretty sure that its absence is accidental.

3. Removed several deprecated and commented-out sections of code.  I can
see no compelling reason for retaining these, given that older versions
can always be retrieved from CVS.

4. Started to remove "// PredictedCdsRetrieval" comments; I'd argue that
if it's really crucial to be able to track which parts of the driver
perform this function then the code should be factored out into its own
class(es).  It not then the CVS commit messages should suffice.  If I'm
missing something here please let me know and I'll revert the changes...

5. Corrected some minor typos (all in comments, not code.)

None of the above changes should have any impact on anyone but developers
working on the chado JDBC adapter.