Fixed Sybase SQL bug.  JdbcChadoAdapter was using "length" to determine the length of a CLOB/TEXT column (feature.residues).  However, Sybase doesn't know anything about this function and needs to use "char_length" instead. I added an abstract method, "getClobLengthFunction", to JdbcChadoAdapter and modified both subclasses to return the appropriate value ("length" for Postgres and "char_length" for Sybase).
Also removed some more commented-out code blocks and changed the return values
of TigrChadoInstance.getGeneNameField and getTranscriptNameField to prevent
another SQL error (which only affected TIGR dbs).  These methods currently return
the column *aliases* of columns that are retrieved/aliased in getCdsSql.
I'll be changing the semantics shortly so that these methods instead return the
actual column names (of columns in the chado feature table); this will essentially
be an implementation of Mark's (?) suggestion to (partially) autogenerate the SQL
for the CDS query.