Slight change (an improvement, hopefully) to the semantics of ChadoInstance.getGeneNameField() and getTranscriptNameField(). Previously these methods returned the names of column aliases that had to appear in the SQL used by JdbcChadoAdapter.addGeneModels. This introduced a somewhat undesirable dependency between the SQL query and any class that implements ChadoInstance.  Now, however, these two methods must return the names of *columns* in the chado feature table.  JdbcChadoAdapter.addGeneModels then takes care of modifying the SQL query accordingly.
That is, the SQL query has a standard set of gene and transcript
columns that it retrieves.  If the requested fields from
getGeneNameField() and getTranscriptNameField() are not among them,
they will be added to the query.  The code then automatically
extracts the correct fields to pass to setName().  The advantage
of this approach is that it's impossible for the get*Field()
methods and the SQL query to get out of sync, although it is
still possible for them to get out of sync with the schema itself.
Hopefully this is an easier condition to check for (and ultimately
it could be done when the chado XML config file is parsed.)

The comments in ChadoInstance.java have been updated accordingly
and I've updated FlybaseChadoInstance and TigrChadoInstance to
ensure backwards-compatibility.