This is a large commit, the main purpose of which is to address a serious problem with the chado adapter's handling of chado analyses.  The current adapter assumes that analysis.program is a unique key for the chado analysis table (i.e., that each program maps to at most one row in the analysis table.)  This is not true in all chado instances (in particular, it is not true of TIGR chado annotation databases, in which each computational analysis typically generates one chado analysis row per contig/assembly.) This assumption is also not supported by the chado index schema, which defines a unique constraint on (program, programversion, sourcename), and has no other unique constraint on any of these three columns.
In response to this problem I've refactored the code somewhat and changed
the semantics of some of the existing classes, all while (hopefully)
maintaining reverse compatibility.  The ChadoProgram class, which was
previously assumed to represent exactly one chado analysis row, now
represents a *set* of chado analysis rows, determined by matching the
program, programversion, and sourcename columns against those stored in
the Java object.  The program must be constrained, but the other two
columns need not be.  In JdbcChadoAdapter the AnalysisTable and AnalysisRow
inner classes from a previous version of the adapter have been reinstated;
AnalysisRow cannot be merged with ChadoProgram (as one of the comments
suggested), for precisely the reason outlined above.

This commit also addresses a secondary problem, namely that an attempt to
use an SQL "IN" clause to retrieve all the chado analysis rows that
correspond to a ChadoProgram or set of ChadoPrograms will fail if there
are too many ids enumerated in the "IN" clause.  Most database servers
have a limit on the size of "IN" clauses, if not on the length of the
entire query.  For now I've added a new adapter configuration parameter,
cacheAnalysisTable.  If set to true (the default, for reverse compatibility):

-The entire analysis table will be cached at startup
-Features linked to a list of ChadoPrograms will be retrieved using
 an SQL "IN" clause.
-Analyis property lookups (for the feature rows returned by
 addOneLevelResults(), for example) will be done using the cache, and the
 SQL for the queries will *not* join through with the analysis table.

If cacheAnalysisTable is set to false then:

-The analysis table will not be cached at startup.
-Features linked to a list of ChadoPrograms will be retrieved using
 an SQL WHERE clause constructed with "AND" and "OR" and using the
 program, programversion, and sourcename values associated with the
 ChadoPrograms.
-Analysis property lookups will be done by constructing SQL queries
 that join through with the analysis table.

Note that the "cacheAnalysisTable" parameter is actually doing a couple
of different things; it's specifying whether to cache the analysis
table, obviously, but it's also dictating whether the feature retrieval
queries should ever join with the analysis table.  One can imagine an
approach in which the analysis table is not cached, but an "IN" query
is still used to retrieve the needed feature rows.   These changes also
don't completely solve the problem of the chado adapter potentially
generating overly-large "IN" clauses; this should be addressed by adding
a new configuration parameter ("max_in_clause_terms"?) and rewriting
the adapter to automatically split 1 large "IN" query into 2 or more
small ones.

Finally, some smaller changes were made:
 ChadoDatabase
  -slight modifications to comment formatting
  -removed unused instance variable
  -removed large chunk of commented-out code that I believe has been superseded by
   ChadoInstance/SeqType
 ChadoInstance/XmlConfigFileParser
  -added 'set/getCacheAnalysisTable' preference, with default = true
  -changed signature of getAnalysisType so that a ResultSet no longer has to be passed in
 JdbcChadoAdapter
  -reinstated AnalysisRow and AnalysisTable; they're not the same as ChadoProgram
  -throw exception if *any* program couldn't be found in analysis table,
   rather than if none of the programs could be found.
  -added getFeatProps, getSynonyms, and getDbXrefs to addOneLevelAnnotations(),
   to match addGeneModels()
 ChadoProgram
  -changed isRetrieveCDS to retrieveCDS
 general
  -fixed some indenting
  -removed large blocks of commented-out code; these can always be retrieved from CVS