Fixed a logic error in addGenePredictionResults().  This method was looking up the predicted CDS and running the following code for each and every exon in the transcript:
   transcriptFeatSet.setTranslationStart(((ChadoCds) predCds.firstElement()).getStart(), false);
   transcriptFeatSet.setTranslationEnd(((ChadoCds) predCds.firstElement()).getEnd());

However, these statements only need to be executed once, preferably *after* all the
exons in the transcript have been retrieved.  I didn't dig deep enough in the code
to see whether the exact timing of the call matters, but it stands to reason that
it would be safer to call setTranslationStart() and setTranslationEnd() on values
that fall inside an exon of the transcript.  For example, if a gene has 3 exons and
exon #1 goes from 100-200, then an attempt to set the translation start to 300--which
falls within exon #2, let's say--may well generate an error if it is done before
exon #2 is added to the transcriptFeatSet.  In the TIGR databases I was seeing a
large number of 'Unable to set translation start' errors from Apollo, and changing
the code to set the translation start and end *once*, after all the exons in a
transcript have been loaded, fixes this.  The other, more minor, problem with the
existing method was the duplication of code required to handle the last transcript
processed by the while loop.  I've factored out the repeated code into a new "helper"
method, setTranscriptNameAndCdsBounds.

I've also modified addGenePredictionResults to use the ChadoInstance
to set the translation start and stop for predicted transcripts--on the grounds that
predicted transcripts should be handled in the same way as annotated transcripts and,
more importantly, because delegating this task allows different adapters/instances to
implement different policies.  In particular, the TIGR version of the adapter may need
to apply a +2/-2 correction to account for the fact that all TIGR CDS features include
the stop codon.  Added the following method to ChadoInstance to support this:

    public void setTranslationStartAndStop(FeatureSet trans, ChadoCds cds);

In order to maintain backwards compatibility the FlybaseChadoInstance implementation of
this method explicitly sets the translation end (which is what the original code was
doing in addGenePredictionResults().)

I've also removed a debug println that was always printing, regardless of the value in
Config.DEBUG.  It looks like this line made it into the 1.6.5 release somehow.