Added --featureName and reverted --location back to its original implementation.  A previous commit had modified --location so that it could be used like this:
1:1000-6000

OR like this (as a shorthand to specify the *entire* coordinate
range of the named feature)

1

However, I've reverted --location back so that it only supports
the first syntax, and always produces the DataInputType BASEPAIR_RANGE.
Instead I've added a --featureName argument that can be used in
conjunction with --inputType to specify that (the entirety of) a
particular feature should be loaded.

Also removed a number of debug printlns that made it into the previous
commit.