Removed hard-coded FlyBase login info. from ChadoAdapterGUI.java.  The comments state that this was "just for 7/2004 release" ;)  However, I've also made some changes to allow logins and passwords to be "hard-coded" in the Apollo configuration file, if so desired.  The <chadodb> XML element now recognizes the following subelements, in addition to <dbUser>:
<dbPassword>
  This value (if present) is passed to ChadoDatabase.setPassword() and used
  to initialize the password text entry field.
<allowDbUserInput>
  Whether the GUI will let the user supply a database username/login.  If
  false then the "Login:" text entry field will not be editable (but will
  still be visible.)
<allowDbPasswordInput>
  Like <allowDbUserInput> but for the database password

These new options are used in the adapter configuration file to emulate
(sort of) the old behavior for one of the FlyBase databases:

	<dbUser>apollo</dbUser>
	<dbPassword>flybase930</dbPassword>
	<!-- Don't allow user to edit username/password: -->
	<allowDbUserInput>false</allowDbUserInput>
	<allowDbPasswordInput>false</allowDbPasswordInput>

Note that there's a comment in the chado-adapter.xml file that says
'user is "apollo" password is "flybase930" for all fly svrs'.  This
contradicts the FlyBase username ("flybase") that was previously hard-
coded in ChadoAdapterGUI.java.  So I'm not sure which one is right
but it seems to be a moot point, since according to Josh Goodman
the public FlyBase dbs at flybase.org were taken down due to security
concerns.  So there you have it.  I've also modified the GUI code so
that the login + password section always appear, but may be disable.
If this becomes a problem (maybe because you want the user to login but
don't want them to *know* the username and password they're using?) then
another option can be added to the chado adapter xml config.